+++ src/StringTable.ts
@@ -179,6 +179,11 @@ export const view_str$compile$cpuType = [
     'CPU Type'
 ][langIndex];
 
+export const view_str$compile$archExtensions = [
+    'Arch 扩展选项',
+    'Arch Extended Options'
+][langIndex];
+
 export const view_str$compile$cpuVendor = [
     'CPU 厂商',
     'CPU Vendor'
@@ -418,6 +423,11 @@ export const view_str$env_desc$compiler_folder = [
 ][langIndex];
 
 //---------------Other---------------
+
+export const view_str$virual_doc_provider_banner = [
+    `通过修改并保存这个文件来更新项目配置（注意这个文件是临时的，不要使用其他工具打开这个文件）`,
+    `You can modify the configuration by editing and saving this file.`
+][langIndex];
 
 export const view_str$prompt$filesOptionsComment = [
     `##########################################################################################`,